# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "961 Apart Docs",
        "description": "Description to Api",
        "version": "2.0"
    },
    "paths": {
...
        "/api/v1/advantage": {
            "get": {
                "tags": [
                    "Advantage"
                ],
                "summary": "Get all",
                "operationId": "9c7eae3000881e8abba9d640c8b16571",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Success",
                                    "data": [
                                        {
                                            "id": 1,
                                            "title": "",
                                            "description": "",
                                            "photo": null,
                                            "created_at": null,
                                            "updated_at": null
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Advantage"
                ],
                "summary": "Create advantage",
                "operationId": "b66d3799a61f853afe598d34cb30a709",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "square_from",
                                    "square_to",
                                    "number_floors",
                                    "number_apartments",
                                    "number_parking_spaces",
                                    "photo",
                                    "relations",
                                    "address_info",
                                    "layout_info",
                                    "price_info"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "integer"
                                    },
                                    "photo": {
                                        "description": "Upload file",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Success",
                                    "data": {
                                        "title": "",
                                        "description": null,
                                        "photo": "advantages/pnGQWDxHsR15YMBPcfas6mYPFquZZkEGOiRNnw1O.jpg",
                                        "updated_at": "2024-09-01T06:20:03.000000Z",
                                        "created_at": "2024-09-01T06:20:03.000000Z",
                                        "id": 10
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/advantage/{id}": {
            "get": {
                "tags": [
                    "Advantage"
                ],
                "summary": "Retrieve Advantage by Id",
                "operationId": "4800106f1df6bfcb9c3e84c7aeb63ae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Advantage id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Success",
                                    "data": {
                                        "title": "",
                                        "description": null,
                                        "photo": "landscapings/pnGQWDxHsR15YMBPcfas6mYPFquZZkEGOiRNnw1O.jpg",
                                        "updated_at": "2024-09-01T06:20:03.000000Z",
                                        "created_at": "2024-09-01T06:20:03.000000Z",
                                        "id": 10
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Advantage"
                ],
                "summary": "Update Advantage",
                "operationId": "dbed3b23dcf30af0aaa0b936d51b495e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Advantage id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "integer"
                                    },
                                    "photo": {
                                        "description": "Upload file",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Advantage"
                ],
                "summary": "Delete advantage by Id",
                "operationId": "eb8a18c77ecfb89a00b2c37420ff3d7c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Advantage id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
...
}